moves balance hook helpers to balance helper file, cleans up response type for get history hook, cleans up error handling in standalone data hooks
