+++ extension/src/helpers/hooks/useGetBalances.tsx
@@ -1,5 +1,4 @@
 import { useReducer } from "react";
-import { Networks } from "stellar-sdk";
 
 import {
   getAccountBalances,
@@ -17,9 +16,6 @@ import { initialState, reducer } from "helpers/request";
 import { storeBalanceMetricData } from "helpers/metrics";
 import { filterHiddenBalances, sortBalances } from "popup/helpers/account";
 import { AssetType } from "@shared/api/types/account-balance";
-import { isContractId } from "popup/helpers/soroban";
-import { getCanonicalFromAsset } from "helpers/stellar";
-import { getAssetSacAddress } from "@shared/helpers/soroban/token";
 
 export const isGetBalancesError = (
   response: AccountBalances | Error,
@@ -28,46 +24,6 @@ export const isGetBalancesError = (
     return true;
   }
   return false;
-};
-
-export const findAddressBalance = (
-  balances: AssetType[],
-  address: string,
-  network: Networks,
-) => {
-  if (address === "native") {
-    return balances.find(
-      (balance) =>
-        "token" in balance &&
-        "type" in balance.token &&
-        balance.token.type === "native",
-    );
-  }
-  if (isContractId(address)) {
-    // first check for contract ID match, then check for SAC match
-    return balances.find((balance) => {
-      if ("contractId" in balance) {
-        return address === balance.contractId;
-      }
-      if ("token" in balance && "issuer" in balance.token) {
-        const canonical = getCanonicalFromAsset(
-          balance.token.code,
-          balance.token.issuer.key,
-        );
-        const sacAddress = getAssetSacAddress(canonical, network);
-        return sacAddress === address;
-      }
-      return false;
-    });
-  }
-
-  return balances.find((balance) => {
-    const balanceIssuer =
-      "token" in balance && "issuer" in balance.token
-        ? balance.token.issuer.key
-        : "";
-    return balanceIssuer === address;
-  });
 };
 
 export interface AccountBalances {
@@ -131,7 +87,7 @@ function useGetBalances(
       return payload;
     } catch (error) {
       dispatch({ type: "FETCH_DATA_ERROR", payload: error });
-      return new Error(JSON.stringify(error));
+      throw new Error("Failed to fetch balances", { cause: error });
     }
   };
