+++ extension/src/helpers/hooks/useGetHistory.tsx
@@ -5,13 +5,24 @@ import { NetworkDetails } from "@shared/constants/stellar";
 import { ServerApi } from "stellar-sdk/lib/horizon";
 import { initialState, reducer } from "helpers/request";
 
+type HistoryResponse = ServerApi.OperationRecord[];
+
+export const isGetHistoryError = (
+  response: HistoryResponse | Error,
+): response is Error => {
+  if (!("balances" in response)) {
+    return true;
+  }
+  return false;
+};
+
 function useGetHistory(publicKey: string, networkDetails: NetworkDetails) {
   const [state, dispatch] = useReducer(
-    reducer<ServerApi.OperationRecord[], unknown>,
+    reducer<HistoryResponse, unknown>,
     initialState,
   );
 
-  const fetchData = async (): Promise<ServerApi.OperationRecord[]> => {
+  const fetchData = async (): Promise<HistoryResponse | Error> => {
     dispatch({ type: "FETCH_DATA_START" });
     try {
       const data = await getAccountHistory(publicKey, networkDetails);
@@ -19,7 +30,7 @@ function useGetHistory(publicKey: string, networkDetails: NetworkDetails) {
       return data;
     } catch (error) {
       dispatch({ type: "FETCH_DATA_ERROR", payload: error });
-      return [];
+      throw new Error("Failed to fetch history", { cause: error });
     }
   };
