+++ extension/src/popup/components/sendPayment/SendSettings/Settings/hooks/useGetSettingsData.tsx
@@ -8,11 +8,11 @@ import { NetworkDetails } from "@shared/constants/stellar";
 import { initialState, reducer } from "helpers/request";
 import {
   AccountBalances,
-  findAddressBalance,
   isGetBalancesError,
   useGetBalances,
 } from "helpers/hooks/useGetBalances";
 import { buildAndSimulateSoroswapTx } from "popup/helpers/sorobanSwap";
+import { findAddressBalance } from "popup/helpers/balance";
 import { stroopToXlm } from "helpers/stellar";
 import {
   CLASSIC_ASSET_DECIMALS,
